add d.blockFormatDetails for details formatting
In the auto completion the details field contains the function
declaration which was previously formatted in a single line per
declaration. This change adds a setting to format each declarations
parameter on a separate line for increased readability.

However due to vscode not supporting this usecase well (it changes the
order of the lines in the details string in the small view and doesn't
show the big view by default + it doesn't support adding indentation at
the start of the line) and because there is currently not a new
workspace-d release which is needed for this, this is a separate branch
to merge some time in the future.

Without bumping the workspace-d version here, the change will not
compile.