Specify SaveOptions in server capabilities
`textDocument/didSave` is specified by the server capability of
SaveOptions or bool in the TextDocumentSyncOptions. Since this wasn't
getting set, some clients may not send `textDocument/didSave` since the
server never broadcasted this capability.

`includeText` is set to `false` since I didn't see anywhere we use the
text contents when responding to this method.

I've confirmed this fixes a bug in the client `helix` as dub build lints
weren't getting triggered since it wouldn't send the
`textDocument/didSave` message unless this capability was sent.